Fix P1/P2 bugs from code review: FSM animation yield ordering, null manager guards, framesSinceSpawn overflow
- animation_manager.h: move yieldToAnimation() calls to after createAnimation() succeeds, so channels aren't left stuck in ANIMATION state on OOM
- main.cpp: add nullptr guards after notificationMgr/animationMgr allocation
- runner_base.h, rain_base.h: widen framesSinceSpawn from uint16_t to uint32_t (overflow at ~55min caused spawn burst)

+++ src/animation/rain/rain_base.h
@@ -95,7 +95,7 @@ public:
 protected:
     // Raindrop state
     Raindrop raindrops[4][MAX_RAINDROP_SLOTS]; // Per channel
-    uint16_t framesSinceSpawn[4];              // Per channel, for spawn probability
+    uint32_t framesSinceSpawn[4];              // Per channel, for spawn probability
 
     // Update raindrops (aging and spawning) — called by update() and inverted subclasses
     void updateRaindrops()
